muchos cambios

+++ src/app/shared/components/modal-subscribe-alert/modal-subscribe-alert.component.ts
@@ -1,11 +1,17 @@
 import { CommonModule } from '@angular/common';
-import { ChangeDetectionStrategy, Component, computed, inject } from '@angular/core';
+import { ChangeDetectionStrategy, Component, Injectable, computed, inject } from '@angular/core';
 import { MatDialogModule } from '@angular/material/dialog';
 import { MatIconModule } from '@angular/material/icon';
 import { Router, RouterModule } from '@angular/router';
 import { toSignal } from '@angular/core/rxjs-interop';
 import { SectionService, SectionServiceData } from '@services/section.service';
 import { AuthService, AuthServiceData, LinkServiceData, Roles } from '@services/auth.service';
+import { HttpClient } from '@angular/common/http';
+import { Observable } from 'rxjs';
+import { MatSnackBar } from '@angular/material/snack-bar';
+@Injectable({
+  providedIn: 'root'
+})
 @Component({
   selector: 'app-modal-subscribe-alert',
   standalone: true,
@@ -23,8 +29,9 @@ export class ModalSubscribeAlertComponent {
 
   private router = inject(Router);
   private authService = inject(AuthService);
+  private _snackBar = inject(MatSnackBar);
   public linkData = computed<LinkServiceData>(()=> this.authService.linkData());
-  constructor(){}
+  constructor(private http: HttpClient){}
 
   ngAfterViewInit(): void {
     this.authService.getCancelar()
@@ -34,10 +41,35 @@ export class ModalSubscribeAlertComponent {
   // linkData().link?.link 
   exitSubscribe(link:any){
 
-    localStorage.clear();
-    window.open(`${link}`, "_blank");
-    this.router.navigate(['/'])
-    localStorage.setItem('role', 'guest');
+    this.http.get<any>(`${link}`)
+    .subscribe((data)=>{
+      if(data.respenvio==='ok'){
+
+        this._snackBar.open(data.msj, '', {
+          duration:2000,
+          horizontalPosition: 'left',
+          verticalPosition: 'bottom',
+        });
+        localStorage.clear();
+        // window.open(`${link}`, "_blank");
+        this.router.navigate(['/'])
+        localStorage.setItem('role', 'guest');
+
+      }else{
+        this._snackBar.open(data.msj, '', {
+          duration:2000,
+          horizontalPosition: 'left',
+          verticalPosition: 'bottom',
+        });
+      }
+  
+
+    })
+
+    // localStorage.clear();
+    // window.open(`${link}`, "_blank");
+    // this.router.navigate(['/'])
+    // localStorage.setItem('role', 'guest');
   }
 
   openOperator(){
